backend/classification-service/requirements.txt
Flask==1.0.2
flask-consulate==0.2.0
flask-restplus==0.12.1
flask-cors==3.0.7

vk==2.0.2
kafka-python==1.4.4
APScheduler==3.5.3
dynaconf==1.2.1

requests==2.21.0

nltk==3.4
numpy==1.16.2
scikit-learn==0.20.3
gensim==3.7.1
pattern==3.6